Pre-load more sbt dependencies
Replace empty sbt project build with a step that uses a snapshot version
of Skink (from skink.tar.gz) to run sbt at Docker build time. This means
that most of the dependencies needed to run sbt nad build Skink will
already be present in the image. This speeds up the first use of sbt
in a container based on the image considerably.