Move to skinklike project
Instead of using real skink as the project to build at Docker build
time, we use a skinklike one. The skinklike project contains the same
dependencies as Skink but only minimal code.

The main reason is that doing the former would require us to put the
Skink code in this repo since Docker build can't access higher levels.
Also, the skinklike project serves the purpose and takes less time to
build, test, etc.